Refactored name of test package
