+++ src/test/java/ServiceTests/Common/HeaderServiceTest.java
@@ -1,4 +1,4 @@
-package ServiceTests.CommonServer;
+package ServiceTests.Common;
 
 import cz.inspire.common.entity.HeaderEntity;
 import cz.inspire.common.repository.HeaderRepository;
