+++ src/test/java/ServiceTests/Common/MenaServiceTest.java
@@ -1,4 +1,4 @@
-package ServiceTests.CommonServer;
+package ServiceTests.Common;
 
 import cz.inspire.common.entity.MenaEntity;
 import cz.inspire.common.repository.MenaRepository;
